+++ src/s_sound.c
@@ -285,7 +285,7 @@ void S_StopSounds(void)
 
 static int S_GetMusicNum(void)
 {
-    static int mnum;
+    int mnum;
 
     if (gamemode == commercial)
     {
@@ -304,17 +304,18 @@ static int S_GetMusicNum(void)
                 mus_ddtblu
             };
 
-            mnum = nmus[(s_randommusic ? M_RandomIntNoRepeat(1, 9, mnum) : gamemap) - 1];
+            mnum = nmus[(s_randommusic ? M_RandomIntNoRepeat(1, 9, gamemap) : gamemap) - 1];
         }
         else
-            mnum = mus_runnin + (s_randommusic ? M_RandomIntNoRepeat(1, 32, mnum) : gamemap) - 1;
+            mnum = mus_runnin + (s_randommusic ? M_RandomIntNoRepeat(1, 32, gamemap) : gamemap) - 1;
     }
     else
     {
         if (gameepisode < 4)
-            mnum = mus_e1m1 + (s_randommusic ? M_RandomIntNoRepeat(1, 21, mnum) : (gameepisode - 1) * 9 + gamemap) - 1;
+            mnum = mus_e1m1 + (s_randommusic ? M_RandomIntNoRepeat(1, 21, (gameepisode - 1) * 9 + gamemap) :
+                (gameepisode - 1) * 9 + gamemap) - 1;
         else if (gameepisode == 5 && sigil)
-            mnum = mus_e5m1 + (s_randommusic ? M_RandomIntNoRepeat(1, 9, mnum) : gamemap) - 1;
+            mnum = mus_e5m1 + (s_randommusic ? M_RandomIntNoRepeat(1, 9, gamemap) : gamemap) - 1;
         else
         {
             int spmus[] =
@@ -331,7 +332,7 @@ static int S_GetMusicNum(void)
                 mus_e1m9    // Tim          E4M9
             };
 
-            mnum = spmus[(s_randommusic ? M_RandomIntNoRepeat(1, 9, mnum) : gamemap) - 1];
+            mnum = spmus[(s_randommusic ? M_RandomIntNoRepeat(1, 9, gamemap) : gamemap) - 1];
         }
     }