+++ releasenotes.md
@@ -5,6 +5,7 @@
 * The name of the PWAD is no longer displayed in the automap if the current map has no title.
 * Several improvements have been made to the support of [*Chex Quest*](https://doomwiki.org/wiki/Chex_Quest), [*Freedoom: Phase 1*](http://freedoom.github.io/) and [*Freedoom: Phase 2*](http://freedoom.github.io/).
 * A bug has been fixed whereby the wrong music could be played in some rare instances.
+* The randomness of the music chosen when the `s_randommusic` CVAR is `on` has been improved.
 * A new `license` CCMD has been implemented that will display the [*GNU General Public License*](https://github.com/bradharding/doomretro/wiki/LICENSE) in the default browser.
 
 ---